Audit layout agnostic components
Entire-Checkpoint: d4c479467395
