+++ crates/ars-components/tests/spec_conformance/layout/aspect_ratio.rs
@@ -1,0 +1,8 @@
+use ars_components::layout::aspect_ratio;
+
+use super::*;
+
+#[test]
+fn aspect_ratio_anatomy_matches_spec() {
+    assert_anatomy("aspect-ratio", &[(aspect_ratio::Part::Root, "root")]);
+}
