+++ crates/ars-components/tests/spec_conformance/layout/grid.rs
@@ -1,0 +1,8 @@
+use ars_components::layout::grid;
+
+use super::*;
+
+#[test]
+fn grid_anatomy_matches_spec() {
+    assert_anatomy("grid", &[(grid::Part::Root, "root")]);
+}
